Extend stylelint-config-recommended
stylelint-config-recommended is the recommended shareable config for
stylelint. It turns on all the [possible errors] rules (not opinionated
stylistic rules) within stylelint.

Note that stylelint-config-recommended enables `at-rule-no-unknown`,
which conflicts with Sass features like `@if`. To work around that, we
use the stylelint-scss plugin's `scss/at-rule-no-unknown` linter, which
includes Sass's @-rule features.

[possible errors]: https://github.com/stylelint/stylelint/blob/master/docs/user-guide/rules.md#possible-errors