Start testing against a sample
To keep up to date with dependencies, it's really hard to know we've
done it correctly without some tests and samples to check against.

This provides a sample (from guides), with which we can test against in
both the positive and negative cases, we include a GitHub Actions
configuration too, so tests are automated.

In our sample, we have `text-size` which doesn't exist in our current
version of `stylelint`. We'll remove this rule in future, as for now, we
override it to ensure the tests start from green.

Finally, we add `lint:css` as a script to npm to allow for "self-linting"
and testing the samples.

https://github.com/thoughtbot/guides/tree/main/sass

Co-authored-by: Luke Mitchell <LkeMitchll@users.noreply.github.com>