Add a script to automate building the CHANGELOG
Taken from Administrate, this helps build a Markdown snippet of the
changes since the last known tag, which gets linked to the PR which
introduced it.

In most cases you do need to do a bit of extra work, but the output can
be copy and pasted into the `CHANGELOG.md` file.

https://github.com/thoughtbot/administrate/blob/main/bin/build-changelog